Add tracking layout per screen/desktop/activity
Layout is now tracked per screen/desktop/activity, instead of one single activity across everything. This should address feature request https://github.com/gerritdevriese/kzones/issues/4

Adding support for screen was a bit tricky to tease out because of the difference between the "active screen" (where the mouse is pointing to) and the "active client" (where the window is on), but otherwise pretty straightforward. The only thing to note is that the default layout for all screen is now the first layout, which might be a bit inconvenient if the user is used to have a main layout that is not the first one in the list. Maybe this could be an config option.